pyproject.toml
[project]
name = "ai-testing"
version = "0.1.0"
description = "AI-powered end-to-end testing and automation platform with RAG capabilities"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # FastAPI & Web
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
    
    # LLM & Agents
    "langgraph>=0.2.0",
    "langchain-core>=1.1.0",
    "langchain-openai>=0.2.0",
    
    # Vector DB & Embeddings
    "pinecone-client>=3.0.0",
    
    # Database
    "pymongo>=4.6.0",
    "psycopg2-binary>=2.9.0",
    
    # Document Processing
    "PyPDF2>=3.16.0",
    "python-docx>=0.8.11",
    "fitz>=0.0.1.dev2",
    
    # API & Utils
    "requests>=2.31.0",
    "python-dotenv>=1.2.1",
    
    # Frontend
    "streamlit>=1.52.0",
    "pandas>=2.1.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"
